Added device support from https://github.com/wcbonner/GoveeBTTempLogger
The https://github.com/wcbonner/GoveeBTTempLogger project seems to have
support for a large number of govee temperature/humidity devices, and the
govee-gateway didn't happen to work with the govee device I had.

I wanted mqtt gatewaying of the govee data, so I changed the device logic
in govee-gateway to use that found in the GoveeBTTempLogger.

This seems to work with the few govee devices I have, but I have yet to get
the H5055 Meat Thermometer to work.

This code could use a cleanup for sure.  The model is noted, but not
sent via MQTT, and devices that support multiple readings are not being
relayed through MQTT yet either.  But, it's a start.

There were also some changes required in the makefile to compile and run
on a raspberry pi with debian 11